+++ Q3E/src/main/java/com/n0n3m4/q3e/Q3EControlView.java
@@ -237,6 +237,7 @@ public class Q3EControlView extends GLSurfaceView implements GLSurfaceView.Rende
 
             Q3E.orig_width = w;
             Q3E.orig_height = h;
+            Q3E.activity.SetupGameViewSize(w, h, false);
 
             UiLoader uildr = new UiLoader(this, gl, Q3E.orig_width, Q3E.orig_height, m_portrait);