+++ Q3E/src/main/java/com/n0n3m4/q3e/Q3EPatchResourceManager.java
@@ -145,7 +145,7 @@ public class Q3EPatchResourceManager
                 Q3EGameConstants.GAME_XASH3D,
                 null,
                 "pak/xash3d/extras.pk3",
-                null
+                ""
         );
         resourceList.add(rsc);
 
@@ -156,7 +156,7 @@ public class Q3EPatchResourceManager
                 Q3EGameConstants.GAME_XASH3D,
                 null,
                 "pak/xash3d/cs16client-extras.pk3",
-                null
+                ""
         );
         resourceList.add(rsc);
     }
@@ -170,6 +170,6 @@ public class Q3EPatchResourceManager
                 return rsc.Fetch(context, overwrite, fsgame);
             }
         }
-        throw new RuntimeException("Unexcept patch resource type: " + type);
+        throw new RuntimeException("Unexpected patch resource type: " + type);
     }
 }