+++ Q3E/src/main/java/com/n0n3m4/q3e/Q3EView.java
@@ -80,6 +80,7 @@ class Q3EView extends SurfaceView implements SurfaceHolder.Callback
             mInit = Q3EMain.gameHelper.Start(holder.getSurface(), w, h);
 
             getHolder().setFixedSize(w, h);
+            Q3E.activity.SetupGameViewSize(getWidth(), getHeight(), true);
         }
     }